+++ tests/plugin_test/src/046_unused_exception.php
@@ -1,0 +1,14 @@
+<?php
+call_user_func(function() {
+    $e = null;
+    try {
+        if (rand() % 2 > 0) {
+            throw new RuntimeException('x');
+        } else {
+            throw new InvalidArgumentException('y');
+        }
+    } catch (RuntimeException $e) {
+        echo $e;
+    } catch (InvalidArgumentException $e2) {
+    }
+});